[s3] http_status is now numeric
There is no type mismatch anymore with backend's request log.

The `target_status_code` is sent always (as string).

Most of the time, `target_status_code` will be the string value of
http_status.

If the original http_status_code is neither equal to the string value of
http_status nor the target_status_code, it is sent additionally.